feat: update patches for hypercube and riscv64 (#556)
* point to local slop and patches

* change `.workspace = true` to `= { workspace = true }`

* replace patch-testing program entrypoints with entrypoint64

* partially update patch testing macro for the new sdk

* use new sdk setup method in patch-testing macro

* update guest-verify-programs Cargo.lock again

* console subscriber init in sp1-test-macro

* point bn at local path instead of git, lockfile

* gate console-subscriber behind tokio_unstable flag

* curve25519-dalek execute works, proving doesn't

* begin attempt to get secp256k1 working

* assert messages for pointer misalignment

* k256 patch proves successfully

* simplify 64 bit build logic a little

* curve25519-dalek-ng prove works (no -ng already worked)

* tiny-keccak prove works

* disable build-host so it can be re-enabled properly later

* bigint execute and prove works

* rsa prove and execute works

* p256 execute and prove works

* sha3 execute and prove works

* sha2 execute and prove works

* fix test-artifacts bugs, a stray warning

* disable secp256k1, enable use of bls patch

* add back patch-testing/Cargo.lock

* secp256k1 (both versions) working

* remove expanded bls test left over from debugging

* make `new_unsound` in sdk and use it in patch-testing

* point at remote patches

* point at remote slop and rrs-succinct

* patch sha2 0.10.9

* re-enable build-host crate

* remove embedded alloc from entrypoint 64

* get examples working

* clean up

* remove stray comment

* halve the ec field element arrays in test-artifacts

* fix: multiple subscribers, new k256 patch

* clippy + fmt

* more fmt

* update examples k256 patch

---------

+++ crates/zkvm/lib/src/lib.rs
@@ -128,7 +128,7 @@ extern "C" {
     /// Computes a big integer operation with a modulus.
     pub fn sys_bigint(
         result: *mut [u64; 4],
-        op: u32,
+        op: u64,
         x: *const [u64; 4],
         y: *const [u64; 4],
         modulus: *const [u64; 4],
